backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2

# URL Analysis
requests==2.31.0

# Security
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3

# Simple ML (no compilation needed)
numpy==1.26.1
